Localization - fix native Intl tests file
